+++ packages/tools/src/discolupe/DivContainer.tsx
@@ -1,4 +1,4 @@
-import { DiscoLupeProject } from './data'
+import { DiscoLupeProject } from './src/model'
 
 export function DivContainer(_: DiscoLupeProject, str: string) {
   return <div>{str}</div>
