+++ packages/tools/src/discolupe/src/model.ts
@@ -24,15 +24,3 @@ export const DiscoLupeModel = z.object({
     projects: z.array(DiscoLupeProject),
   }),
 })
-
-export async function fetchData() {
-  const req = await fetch('http://localhost:3000/api/discolupe')
-
-  const result = await req.json()
-  const parsed = DiscoLupeModel.parse(result)
-  if (!parsed.success) {
-    throw new Error('expected it to succeed')
-  }
-
-  return parsed
-}
